+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/utils/HttpClientUtils.java
@@ -15,6 +15,7 @@
  */
 package io.fabric8.kubernetes.client.utils;
 
+import io.fabric8.kubernetes.api.model.ListOptions;
 import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.internal.SSLUtils;
@@ -64,6 +65,43 @@ public class HttpClientUtils {
 
   public static OkHttpClient createHttpClientForMockServer(final Config config) {
       return createHttpClient(config, b -> b.protocols(Collections.singletonList(Protocol.HTTP_1_1)));
+  }
+
+  public static HttpUrl.Builder appendListOptionParams(HttpUrl.Builder urlBuilder, ListOptions listOptions) {
+    if (listOptions == null) {
+      return urlBuilder;
+    }
+    if (listOptions.getLimit() != null) {
+      urlBuilder.addQueryParameter("limit", listOptions.getLimit().toString());
+    }
+    if (listOptions.getContinue() != null) {
+      urlBuilder.addQueryParameter("continue", listOptions.getContinue());
+    }
+
+    if (listOptions.getResourceVersion() != null) {
+      urlBuilder.addQueryParameter("resourceVersion", listOptions.getResourceVersion());
+    }
+
+    if (listOptions.getFieldSelector() != null) {
+      urlBuilder.addQueryParameter("fieldSelector", listOptions.getFieldSelector());
+    }
+
+    if (listOptions.getLabelSelector() != null) {
+      urlBuilder.addQueryParameter("labelSelector", listOptions.getLabelSelector());
+    }
+
+    if (listOptions.getTimeoutSeconds() != null) {
+      urlBuilder.addQueryParameter("timeoutSeconds", listOptions.getTimeoutSeconds().toString());
+    }
+
+    if (listOptions.getAllowWatchBookmarks() != null) {
+      urlBuilder.addQueryParameter("allowWatchBookmarks", listOptions.getAllowWatchBookmarks().toString());
+    }
+
+    if (listOptions.getWatch() != null) {
+      urlBuilder.addQueryParameter("watch", listOptions.getWatch().toString());
+    }
+    return urlBuilder;
   }
 
   private static OkHttpClient createHttpClient(final Config config, final Consumer<OkHttpClient.Builder> additionalConfig) {